refactor: combine dio clients
